user creation is advised
Not strictly needed. However, since the username is common and known to everyone, chances are high that someone will be able to access the instance if he comes under the allowed block of IPs and has the pem/ppk file. May be create a password and disable SSH login for ubuntu, or create a new user and setup keybased/password_based login for that user, just to be on the safer side.